requirements.txt

jupyter
pydot
graphviz
matplotlib
numpy
PyYAML
scikit-learn
tensorflow
tensorflow_datasets
tqdm
qkeras==0.9.0
keras-flops
git+https://github.com/kentaroy47/keras-Opcounter
keras-tuner
setgpu
mplhep
git+https://github.com/anmeza/hls4ml.git@relu_fuse#egg=hls4ml[profiling]